Order cache by tiles
This changes the cache directory layout to give its contents more
semantic information.

It creates sub-directories like `-48+013/` when cache images are saved.

This allows to see which data has been downloaded for a region.
It also enables manual pruning of a region if needed without additional
tools to deal with the Slippy tile names.

TODO: It does not move existing files into the new structure yet.